feat: rework serializer system

+++ src/api/java/dev/tilera/cwg/api/serialize/RepresentationType.java
@@ -2,9 +2,9 @@ package dev.tilera.cwg.api.serialize;
 
 public enum RepresentationType {
     STRING(String.class),
-    INTEGER(Integer.class),
+    INTEGER(Long.class),
     BOOLEAN(Boolean.class),
-    DOUBLE(Double.class),
+    FLOAT(Double.class),
     OBJECT(Object.class),
     ARRAY(Object[].class);
